lsp: Reimplement `textDocument/documentSymbol`
Instead of computing document symbols in the lanaguage server process,
we now read the symbols out of the sphinx agent's database.
