+++ lib/esbonio/esbonio/server/cli.py
@@ -73,6 +73,7 @@ def main(argv: Optional[Sequence[str]] = None):
         "esbonio.server.features.sphinx_manager",
         "esbonio.server.features.preview_manager",
         "esbonio.server.features.sphinx_support.diagnostics",
+        "esbonio.server.features.sphinx_support.symbols",
     ]
 
     for mod in args.included_modules:
