+++ lib/esbonio/tests/e2e/test_e2e_diagnostics.py
@@ -11,84 +11,6 @@ SERVER_CMD = ["-m", "esbonio"]
 TEST_DIR = pathlib.Path(__file__).parent.parent
 
 
-@pytest_lsp.fixture(
-    scope="module",
-    config=ClientServerConfig(
-        server_command=[sys.executable, *SERVER_CMD],
-    ),
-)
-async def pull_client(lsp_client: LanguageClient, uri_for, tmp_path_factory):
-    """A client that supports the pull-diagnostics model."""
-    build_dir = tmp_path_factory.mktemp("build")
-    workspace_uri = uri_for("sphinx-default", "workspace")
-    test_uri = workspace_uri / "definitions.rst"
-
-    await lsp_client.initialize_session(
-        types.InitializeParams(
-            capabilities=types.ClientCapabilities(
-                # Signal pull diagnostic support
-                text_document=types.TextDocumentClientCapabilities(
-                    diagnostic=types.DiagnosticClientCapabilities(
-                        dynamic_registration=False
-                    )
-                ),
-                # Signal workDoneProgress/create support.
-                window=types.WindowClientCapabilities(
-                    work_done_progress=True,
-                ),
-            ),
-            initialization_options={
-                "server": {"logLevel": "debug"},
-                "sphinx": {
-                    "buildCommand": [
-                        "sphinx-build",
-                        "-M",
-                        "html",
-                        workspace_uri.fs_path,
-                        str(build_dir),
-                    ],
-                    "pythonCommand": [sys.executable],
-                },
-            },
-            workspace_folders=[
-                types.WorkspaceFolder(uri=str(workspace_uri), name="sphinx-default"),
-            ],
-        )
-    )
-
-    # Open a text document to trigger sphinx client creation.
-    lsp_client.text_document_did_open(
-        types.DidOpenTextDocumentParams(
-            text_document=types.TextDocumentItem(
-                uri=str(test_uri),
-                language_id="restructuredtext",
-                version=0,
-                text=pathlib.Path(test_uri).read_text(),
-            )
-        )
-    )
-
-    await lsp_client.wait_for_notification("sphinx/appCreated")
-
-    # Save the document to trigger a build
-    lsp_client.text_document_did_save(
-        types.DidSaveTextDocumentParams(
-            text_document=types.TextDocumentIdentifier(uri=str(test_uri))
-        )
-    )
-
-    build_finished = False
-    while not build_finished:
-        await lsp_client.wait_for_notification("$/progress")
-        report = list(lsp_client.progress_reports.values())[0][-1]
-        build_finished = report.kind == "end"
-
-    yield
-
-    # Teardown
-    await lsp_client.shutdown_session()
-
-
 @pytest.mark.asyncio
 @pytest.mark.skip
 async def test_document_diagnostic(pull_client: LanguageClient, uri_for):
@@ -96,7 +18,7 @@ async def test_document_diagnostic(pull_client: LanguageClient, uri_for):
 
     workspace_uri = uri_for("sphinx-default", "workspace")
     test_uri = workspace_uri / "definitions.rst"
-    report = await pull_client.text_document_diagnostic_async(
+    report = await client.text_document_diagnostic_async(
         types.DocumentDiagnosticParams(
             text_document=types.TextDocumentIdentifier(uri=str(test_uri))
         )
@@ -112,14 +34,14 @@ async def test_document_diagnostic(pull_client: LanguageClient, uri_for):
         "unknown document: '/changelog'",
     }
 
-    assert len(pull_client.diagnostics) == 0, "Server should not publish diagnostics"
+    assert len(client.diagnostics) == 0, "Server should not publish diagnostics"
 
 
 @pytest.mark.asyncio
 @pytest.mark.skip
 async def test_workspace_diagnostic(pull_client: LanguageClient, uri_for):
     """Ensure that we can get diagnostics for the whole workspace correctly."""
-    report = await pull_client.workspace_diagnostic_async(
+    report = await client.workspace_diagnostic_async(
         types.WorkspaceDiagnosticParams(previous_result_ids=[])
     )
 
@@ -138,7 +60,7 @@ async def test_workspace_diagnostic(pull_client: LanguageClient, uri_for):
     for item in report.items:
         assert expected[item.uri] == {d.message for d in item.items}
 
-    assert len(pull_client.diagnostics) == 0, "Server should not publish diagnostics"
+    assert len(client.diagnostics) == 0, "Server should not publish diagnostics"
 
 
 @pytest_lsp.fixture(
